+++ src/main/java/org/guanzon/cas/clients/resultSet2XML/ModelClientAddress.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package org.guanzon.clients.resultSet2XML;
+package org.guanzon.cas.clients.resultSet2XML;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
