+++ src/main/java/org/guanzon/cas/clients/resultSet2XML/ModelClientContactPerson.java
@@ -1,11 +1,11 @@
-package org.guanzon.clients.resultSet2XML;
+package org.guanzon.cas.clients.resultSet2XML;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import org.guanzon.appdriver.base.GRider;
 import org.guanzon.appdriver.base.MiscUtil;
 
-public class ModelClientMobile {
+public class ModelClientContactPerson {
     public static void main (String [] args){
         String path;
         if(System.getProperty("os.name").toLowerCase().contains("win")){
@@ -25,39 +25,32 @@ public class ModelClientMobile {
 
         System.out.println("Connected");
         
-        System.setProperty("sys.default.path.metadata", "D:/GGC_Maven_Systems/config/metadata/Model_Client_Mobile.xml");
+        System.setProperty("sys.default.path.metadata", "D:/GGC_Maven_Systems/config/metadata/Model_Client_Institution_Contact.xml");
         
         
         String lsSQL = "SELECT" +
-                    "  sMobileID" +
+                    " sContctID" +
                     ", sClientID" +
+                    ", sCPerson1" +
+                    ", sCPPosit1" +
                     ", sMobileNo" +
-                    ", cMobileTp" +
-                    ", cOwnerxxx" +
+                    ", sTelNoxxx" +
+                    ", sFaxNoxxx" +
+                    ", sEMailAdd" +
+                    ", sAccount1" +
+                    ", sAccount2" +
+                    ", sAccount3" +
+                    ", sRemarksx" +
                     ", cPrimaryx" +
-                    ", cIncdMktg" +
-                    ", nUnreachx" +
-                    ", dLastVeri" +
-                    ", dInactive" +
-                    ", nNoRetryx" +
-                    ", cInvalidx" +
-                    ", cConfirmd" +
-                    ", dConfirmd" +
-                    ", cSubscrbr" +
-                    ", dHoldMktg" +
-                    ", dMktgMsg1" +
-                    ", dMktgMsg2" +
-                    ", dMktgMsg3" +
-                    ", cNewMobil" +
                     ", cRecdStat" +
                     ", dModified" +
-                        " FROM Client_Mobile" +
+                        " FROM Client_Institution_Contact_Person" +
                         " WHERE 0=1";
         
         
         ResultSet loRS = instance.executeQuery(lsSQL);
         try {
-            if (MiscUtil.resultSet2XML(instance, loRS, System.getProperty("sys.default.path.metadata"), "Client_Mobile", "")){
+            if (MiscUtil.resultSet2XML(instance, loRS, System.getProperty("sys.default.path.metadata"), "Client_Institution_Contact_Person", "")){
                 System.out.println("ResultSet exported.");
             }
         } catch (SQLException e) {
