+++ src/main/java/org/guanzon/cas/clients/resultSet2XML/ModelClientMobile.java
@@ -1,11 +1,11 @@
-package org.guanzon.clients.resultSet2XML;
+package org.guanzon.cas.clients.resultSet2XML;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import org.guanzon.appdriver.base.GRider;
 import org.guanzon.appdriver.base.MiscUtil;
 
-public class ModelClientMail {
+public class ModelClientMobile {
     public static void main (String [] args){
         String path;
         if(System.getProperty("os.name").toLowerCase().contains("win")){
@@ -25,24 +25,39 @@ public class ModelClientMail {
 
         System.out.println("Connected");
         
-        System.setProperty("sys.default.path.metadata", "D:/GGC_Maven_Systems/config/metadata/Model_Client_eMail.xml");
+        System.setProperty("sys.default.path.metadata", "D:/GGC_Maven_Systems/config/metadata/Model_Client_Mobile.xml");
         
         
         String lsSQL = "SELECT" +
-                    "  sEmailIDx" +
+                    "  sMobileID" +
                     ", sClientID" +
-                    ", sEMailAdd" +
+                    ", sMobileNo" +
+                    ", cMobileTp" +
                     ", cOwnerxxx" +
                     ", cPrimaryx" +
+                    ", cIncdMktg" +
+                    ", nUnreachx" +
+                    ", dLastVeri" +
+                    ", dInactive" +
+                    ", nNoRetryx" +
+                    ", cInvalidx" +
+                    ", cConfirmd" +
+                    ", dConfirmd" +
+                    ", cSubscrbr" +
+                    ", dHoldMktg" +
+                    ", dMktgMsg1" +
+                    ", dMktgMsg2" +
+                    ", dMktgMsg3" +
+                    ", cNewMobil" +
                     ", cRecdStat" +
                     ", dModified" +
-                        " FROM Client_eMail_Address" +
+                        " FROM Client_Mobile" +
                         " WHERE 0=1";
         
         
         ResultSet loRS = instance.executeQuery(lsSQL);
         try {
-            if (MiscUtil.resultSet2XML(instance, loRS, System.getProperty("sys.default.path.metadata"), "Client_eMail_Address", "")){
+            if (MiscUtil.resultSet2XML(instance, loRS, System.getProperty("sys.default.path.metadata"), "Client_Mobile", "")){
                 System.out.println("ResultSet exported.");
             }
         } catch (SQLException e) {
